Move UTC types to an external file
