+++ src/services/specialized-devices/switch-manager.interface.ts
@@ -1,14 +1,8 @@
+import { UtcOffset } from '@app/types/datetime.types'
 import { ModuleQuery } from '@app/types/query-common.types'
 
 export type SwitchState = 'ON' | 'OFF'
-
-type SingleDigit = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9
-
-export type ScheduleUtcOffset =
-  | `+${SingleDigit}${SingleDigit}:${SingleDigit}${SingleDigit}`
-  | `+${SingleDigit}`
-  | `-${SingleDigit}${SingleDigit}:${SingleDigit}${SingleDigit}`
-  | `-${SingleDigit}`
+export type ScheduleUtcOffset = UtcOffset
 
 export interface BaseSchedule {
   utcOffset: ScheduleUtcOffset
